refactor(api): fix lint and unit test errors
