+++ packages/api-client/__tests__/api/getFacet.spec.ts
@@ -1,12 +1,10 @@
 import getFacet from '../../src/api/getFacet';
-import defaultQuery from '../../src/api/getFacet/defaultQuery';
+import defaultQuery from '../../src/api/getFacet/searchQuery';
 import { Context } from '../../src/types';
 
 describe('[vendure-api-client] getFacet', () => {
   it('fetches products, collections, and facets for the current query', async () => {
-    const givenVariables = {
-      input: {}
-    };
+    const givenVariables = {};
 
     const context = {
       config: {},
@@ -21,7 +19,7 @@ describe('[vendure-api-client] getFacet', () => {
       extendQuery: (customQuery, args) => args
     } as unknown as Context;
 
-    const { data } = await getFacet(context, { input: {}});
+    const { data } = await getFacet(context, {});
 
     expect(data).toBe('facet response');
   });
