+++ packages/api-client/src/helpers/apiClient/defaultSettings.ts
@@ -1,9 +1,9 @@
-import { ClientConfig } from "../../types/setup";
+import { ClientConfig } from '../../types/setup';
 
 export const defaultSettings: ClientConfig = {
   api: {
     uri: ''
   },
   currency: 'USD',
-  lang: 'en',
-}
+  lang: 'en'
+};
