+++ packages/api-client/src/helpers/vendureLink/graphQL.ts
@@ -3,23 +3,23 @@ import ApolloClient from 'apollo-client';
 import fetch from 'isomorphic-fetch';
 import { ApolloLink } from 'apollo-link';
 import { createHttpLink } from 'apollo-link-http';
-import { InMemoryCache } from 'apollo-cache-inmemory';
+import { InMemoryCache, NormalizedCacheObject } from 'apollo-cache-inmemory';
 import { Logger } from '@vue-storefront/core';
 import { onError } from 'apollo-link-error';
 import { RetryLink } from 'apollo-link-retry';
-import { setContext } from 'apollo-link-context';
+// import { setContext } from 'apollo-link-context';
 import { handleRetry } from './linkHandlers';
 import { Config } from '../../types/setup';
 
 const createErrorHandler = () => onError(({
   graphQLErrors,
-  networkError,
+  networkError
 }) => {
   if (graphQLErrors) {
     graphQLErrors.forEach(({
       message,
       locations,
-      path,
+      path
     }) => {
       if (!message.includes('Resource Owner Password Credentials Grant')) {
         if (!locations) {
@@ -29,7 +29,7 @@ const createErrorHandler = () => onError(({
 
         const parsedLocations = locations.map(({
           column,
-          line,
+          line
         }) => `[column: ${column}, line: ${line}]`);
 
         Logger.error(`[GraphQL error]: Message: ${message}, Location: ${parsedLocations.join(', ')}, Path: ${path}`);
@@ -43,11 +43,11 @@ const createErrorHandler = () => onError(({
 });
 
 export const apolloLinkFactory = (
-  settings: Config, 
-  handlers?: {
-    authLink?: ApolloLink;
-  }
-) => {
+  settings: Config
+  // handlers?: {
+  // authLink?: ApolloLink;
+  // }
+): ApolloLink => {
   // TODO: Configure later when token management will be needed
 //   const baseAuthLink = handlers?.authLink || setContext((apolloReq, { headers }) => ({
 //     headers: {
@@ -60,10 +60,9 @@ export const apolloLinkFactory = (
 
   const errorRetry = new RetryLink({
     attempts: handleRetry(),
-    delay: () => 0,
+    delay: () => 0
   });
 
-  // eslint-disable-next-line unicorn/prefer-spread
   return ApolloLink.from([
     onErrorLink,
     errorRetry,
@@ -72,7 +71,7 @@ export const apolloLinkFactory = (
   ]);
 };
 
-export const apolloClientFactory = (customOptions: Record<string, any>) => new ApolloClient({
+export const apolloClientFactory = (customOptions: Record<string, any>): ApolloClient<NormalizedCacheObject> => new ApolloClient({
   cache: new InMemoryCache(),
-  ...customOptions,
+  ...customOptions
 });
