+++ packages/api-client/src/types/API.ts
@@ -1,4 +1,4 @@
-import { getProduct, getCategory, getFacet } from "../api";
+import { getProduct, getCategory, getFacet } from '../api';
 
 export interface VendureApiMethods {
   getProduct: typeof getProduct;
