+++ packages/api-client/src/types/GraphQL.ts
@@ -2,6 +2,7 @@ export type Maybe<T> = T | null;
 export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
 export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
 export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
+
 /** All built-in and custom scalars, mapped to their actual values */
 export type Scalars = {
   ID: string;
@@ -9,10 +10,13 @@ export type Scalars = {
   Boolean: boolean;
   Int: number;
   Float: number;
+
   /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
   DateTime: any;
+
   /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
   JSON: any;
+
   /** The `Upload` scalar type represents a file upload. */
   Upload: any;
 };
@@ -156,7 +160,6 @@ export type Collection = Node & {
   customFields?: Maybe<Scalars['JSON']>;
 };
 
-
 export type CollectionProductVariantsArgs = {
   options?: Maybe<ProductVariantListOptions>;
 };
@@ -242,6 +245,7 @@ export type ConfigArgDefinition = {
 
 export type ConfigArgInput = {
   name: Scalars['String'];
+
   /** A JSON stringified representation of the actual value */
   value: Scalars['String'];
 };
@@ -353,318 +357,475 @@ export type CreateCustomerInput = {
  * @docsCategory common
  */
 export enum CurrencyCode {
+
   /** United Arab Emirates dirham */
   Aed = 'AED',
+
   /** Afghan afghani */
   Afn = 'AFN',
+
   /** Albanian lek */
   All = 'ALL',
+
   /** Armenian dram */
   Amd = 'AMD',
+
   /** Netherlands Antillean guilder */
   Ang = 'ANG',
+
   /** Angolan kwanza */
   Aoa = 'AOA',
+
   /** Argentine peso */
   Ars = 'ARS',
+
   /** Australian dollar */
   Aud = 'AUD',
+
   /** Aruban florin */
   Awg = 'AWG',
+
   /** Azerbaijani manat */
   Azn = 'AZN',
+
   /** Bosnia and Herzegovina convertible mark */
   Bam = 'BAM',
+
   /** Barbados dollar */
   Bbd = 'BBD',
+
   /** Bangladeshi taka */
   Bdt = 'BDT',
+
   /** Bulgarian lev */
   Bgn = 'BGN',
+
   /** Bahraini dinar */
   Bhd = 'BHD',
+
   /** Burundian franc */
   Bif = 'BIF',
+
   /** Bermudian dollar */
   Bmd = 'BMD',
+
   /** Brunei dollar */
   Bnd = 'BND',
+
   /** Boliviano */
   Bob = 'BOB',
+
   /** Brazilian real */
   Brl = 'BRL',
+
   /** Bahamian dollar */
   Bsd = 'BSD',
+
   /** Bhutanese ngultrum */
   Btn = 'BTN',
+
   /** Botswana pula */
   Bwp = 'BWP',
+
   /** Belarusian ruble */
   Byn = 'BYN',
+
   /** Belize dollar */
   Bzd = 'BZD',
+
   /** Canadian dollar */
   Cad = 'CAD',
+
   /** Congolese franc */
   Cdf = 'CDF',
+
   /** Swiss franc */
   Chf = 'CHF',
+
   /** Chilean peso */
   Clp = 'CLP',
+
   /** Renminbi (Chinese) yuan */
   Cny = 'CNY',
+
   /** Colombian peso */
   Cop = 'COP',
+
   /** Costa Rican colon */
   Crc = 'CRC',
+
   /** Cuban convertible peso */
   Cuc = 'CUC',
+
   /** Cuban peso */
   Cup = 'CUP',
+
   /** Cape Verde escudo */
   Cve = 'CVE',
+
   /** Czech koruna */
   Czk = 'CZK',
+
   /** Djiboutian franc */
   Djf = 'DJF',
+
   /** Danish krone */
   Dkk = 'DKK',
+
   /** Dominican peso */
   Dop = 'DOP',
+
   /** Algerian dinar */
   Dzd = 'DZD',
+
   /** Egyptian pound */
   Egp = 'EGP',
+
   /** Eritrean nakfa */
   Ern = 'ERN',
+
   /** Ethiopian birr */
   Etb = 'ETB',
+
   /** Euro */
   Eur = 'EUR',
+
   /** Fiji dollar */
   Fjd = 'FJD',
+
   /** Falkland Islands pound */
   Fkp = 'FKP',
+
   /** Pound sterling */
   Gbp = 'GBP',
+
   /** Georgian lari */
   Gel = 'GEL',
+
   /** Ghanaian cedi */
   Ghs = 'GHS',
+
   /** Gibraltar pound */
   Gip = 'GIP',
+
   /** Gambian dalasi */
   Gmd = 'GMD',
+
   /** Guinean franc */
   Gnf = 'GNF',
+
   /** Guatemalan quetzal */
   Gtq = 'GTQ',
+
   /** Guyanese dollar */
   Gyd = 'GYD',
+
   /** Hong Kong dollar */
   Hkd = 'HKD',
+
   /** Honduran lempira */
   Hnl = 'HNL',
+
   /** Croatian kuna */
   Hrk = 'HRK',
+
   /** Haitian gourde */
   Htg = 'HTG',
+
   /** Hungarian forint */
   Huf = 'HUF',
+
   /** Indonesian rupiah */
   Idr = 'IDR',
+
   /** Israeli new shekel */
   Ils = 'ILS',
+
   /** Indian rupee */
   Inr = 'INR',
+
   /** Iraqi dinar */
   Iqd = 'IQD',
+
   /** Iranian rial */
   Irr = 'IRR',
+
   /** Icelandic króna */
   Isk = 'ISK',
+
   /** Jamaican dollar */
   Jmd = 'JMD',
+
   /** Jordanian dinar */
   Jod = 'JOD',
+
   /** Japanese yen */
   Jpy = 'JPY',
+
   /** Kenyan shilling */
   Kes = 'KES',
+
   /** Kyrgyzstani som */
   Kgs = 'KGS',
+
   /** Cambodian riel */
   Khr = 'KHR',
+
   /** Comoro franc */
   Kmf = 'KMF',
+
   /** North Korean won */
   Kpw = 'KPW',
+
   /** South Korean won */
   Krw = 'KRW',
+
   /** Kuwaiti dinar */
   Kwd = 'KWD',
+
   /** Cayman Islands dollar */
   Kyd = 'KYD',
+
   /** Kazakhstani tenge */
   Kzt = 'KZT',
+
   /** Lao kip */
   Lak = 'LAK',
+
   /** Lebanese pound */
   Lbp = 'LBP',
+
   /** Sri Lankan rupee */
   Lkr = 'LKR',
+
   /** Liberian dollar */
   Lrd = 'LRD',
+
   /** Lesotho loti */
   Lsl = 'LSL',
+
   /** Libyan dinar */
   Lyd = 'LYD',
+
   /** Moroccan dirham */
   Mad = 'MAD',
+
   /** Moldovan leu */
   Mdl = 'MDL',
+
   /** Malagasy ariary */
   Mga = 'MGA',
+
   /** Macedonian denar */
   Mkd = 'MKD',
+
   /** Myanmar kyat */
   Mmk = 'MMK',
+
   /** Mongolian tögrög */
   Mnt = 'MNT',
+
   /** Macanese pataca */
   Mop = 'MOP',
+
   /** Mauritanian ouguiya */
   Mru = 'MRU',
+
   /** Mauritian rupee */
   Mur = 'MUR',
+
   /** Maldivian rufiyaa */
   Mvr = 'MVR',
+
   /** Malawian kwacha */
   Mwk = 'MWK',
+
   /** Mexican peso */
   Mxn = 'MXN',
+
   /** Malaysian ringgit */
   Myr = 'MYR',
+
   /** Mozambican metical */
   Mzn = 'MZN',
+
   /** Namibian dollar */
   Nad = 'NAD',
+
   /** Nigerian naira */
   Ngn = 'NGN',
+
   /** Nicaraguan córdoba */
   Nio = 'NIO',
+
   /** Norwegian krone */
   Nok = 'NOK',
+
   /** Nepalese rupee */
   Npr = 'NPR',
+
   /** New Zealand dollar */
   Nzd = 'NZD',
+
   /** Omani rial */
   Omr = 'OMR',
+
   /** Panamanian balboa */
   Pab = 'PAB',
+
   /** Peruvian sol */
   Pen = 'PEN',
+
   /** Papua New Guinean kina */
   Pgk = 'PGK',
+
   /** Philippine peso */
   Php = 'PHP',
+
   /** Pakistani rupee */
   Pkr = 'PKR',
+
   /** Polish złoty */
   Pln = 'PLN',
+
   /** Paraguayan guaraní */
   Pyg = 'PYG',
+
   /** Qatari riyal */
   Qar = 'QAR',
+
   /** Romanian leu */
   Ron = 'RON',
+
   /** Serbian dinar */
   Rsd = 'RSD',
+
   /** Russian ruble */
   Rub = 'RUB',
+
   /** Rwandan franc */
   Rwf = 'RWF',
+
   /** Saudi riyal */
   Sar = 'SAR',
+
   /** Solomon Islands dollar */
   Sbd = 'SBD',
+
   /** Seychelles rupee */
   Scr = 'SCR',
+
   /** Sudanese pound */
   Sdg = 'SDG',
+
   /** Swedish krona/kronor */
   Sek = 'SEK',
+
   /** Singapore dollar */
   Sgd = 'SGD',
+
   /** Saint Helena pound */
   Shp = 'SHP',
+
   /** Sierra Leonean leone */
   Sll = 'SLL',
+
   /** Somali shilling */
   Sos = 'SOS',
+
   /** Surinamese dollar */
   Srd = 'SRD',
+
   /** South Sudanese pound */
   Ssp = 'SSP',
+
   /** São Tomé and Príncipe dobra */
   Stn = 'STN',
+
   /** Salvadoran colón */
   Svc = 'SVC',
+
   /** Syrian pound */
   Syp = 'SYP',
+
   /** Swazi lilangeni */
   Szl = 'SZL',
+
   /** Thai baht */
   Thb = 'THB',
+
   /** Tajikistani somoni */
   Tjs = 'TJS',
+
   /** Turkmenistan manat */
   Tmt = 'TMT',
+
   /** Tunisian dinar */
   Tnd = 'TND',
+
   /** Tongan paʻanga */
   Top = 'TOP',
+
   /** Turkish lira */
   Try = 'TRY',
+
   /** Trinidad and Tobago dollar */
   Ttd = 'TTD',
+
   /** New Taiwan dollar */
   Twd = 'TWD',
+
   /** Tanzanian shilling */
   Tzs = 'TZS',
+
   /** Ukrainian hryvnia */
   Uah = 'UAH',
+
   /** Ugandan shilling */
   Ugx = 'UGX',
+
   /** United States dollar */
   Usd = 'USD',
+
   /** Uruguayan peso */
   Uyu = 'UYU',
+
   /** Uzbekistan som */
   Uzs = 'UZS',
+
   /** Venezuelan bolívar soberano */
   Ves = 'VES',
+
   /** Vietnamese đồng */
   Vnd = 'VND',
+
   /** Vanuatu vatu */
   Vuv = 'VUV',
+
   /** Samoan tala */
   Wst = 'WST',
+
   /** CFA franc BEAC */
   Xaf = 'XAF',
+
   /** East Caribbean dollar */
   Xcd = 'XCD',
+
   /** CFA franc BCEAO */
   Xof = 'XOF',
+
   /** CFP franc (franc Pacifique) */
   Xpf = 'XPF',
+
   /** Yemeni rial */
   Yer = 'YER',
+
   /** South African rand */
   Zar = 'ZAR',
+
   /** Zambian kwacha */
   Zmw = 'ZMW',
+
   /** Zimbabwean dollar */
   Zwl = 'ZWL'
 }
@@ -712,7 +873,6 @@ export type Customer = Node & {
   customFields?: Maybe<Scalars['JSON']>;
 };
 
-
 export type CustomerOrdersArgs = {
   options?: Maybe<OrderListOptions>;
 };
@@ -735,7 +895,6 @@ export type CustomerGroup = Node & {
   name: Scalars['String'];
   customers: CustomerList;
 };
-
 
 export type CustomerGroupCustomersArgs = {
   options?: Maybe<CustomerListOptions>;
@@ -777,7 +936,6 @@ export type DateRange = {
   end: Scalars['DateTime'];
 };
 
-
 /**
  * Expects the same validation formats as the `<input type="datetime-local">` HTML element.
  * See https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/datetime-local#Additional_attributes
@@ -803,8 +961,10 @@ export type DeletionResponse = {
 };
 
 export enum DeletionResult {
+
   /** The entity was successfully deleted */
   Deleted = 'DELETED',
+
   /** Deletion did not take place, reason given in message */
   NotDeleted = 'NOT_DELETED'
 }
@@ -1093,7 +1253,6 @@ export type InvalidCredentialsError = ErrorResult & {
   authenticationError: Scalars['String'];
 };
 
-
 /**
  * @description
  * Languages in the form of a ISO 639-1 language code with optional
@@ -1104,318 +1263,475 @@ export type InvalidCredentialsError = ErrorResult & {
  * @docsCategory common
  */
 export enum LanguageCode {
+
   /** Afrikaans */
   Af = 'af',
+
   /** Akan */
   Ak = 'ak',
+
   /** Albanian */
   Sq = 'sq',
+
   /** Amharic */
   Am = 'am',
+
   /** Arabic */
   Ar = 'ar',
+
   /** Armenian */
   Hy = 'hy',
+
   /** Assamese */
   As = 'as',
+
   /** Azerbaijani */
   Az = 'az',
+
   /** Bambara */
   Bm = 'bm',
+
   /** Bangla */
   Bn = 'bn',
+
   /** Basque */
   Eu = 'eu',
+
   /** Belarusian */
   Be = 'be',
+
   /** Bosnian */
   Bs = 'bs',
+
   /** Breton */
   Br = 'br',
+
   /** Bulgarian */
   Bg = 'bg',
+
   /** Burmese */
   My = 'my',
+
   /** Catalan */
   Ca = 'ca',
+
   /** Chechen */
   Ce = 'ce',
+
   /** Chinese */
   Zh = 'zh',
+
   /** Simplified Chinese */
   ZhHans = 'zh_Hans',
+
   /** Traditional Chinese */
   ZhHant = 'zh_Hant',
+
   /** Church Slavic */
   Cu = 'cu',
+
   /** Cornish */
   Kw = 'kw',
+
   /** Corsican */
   Co = 'co',
+
   /** Croatian */
   Hr = 'hr',
+
   /** Czech */
   Cs = 'cs',
+
   /** Danish */
   Da = 'da',
+
   /** Dutch */
   Nl = 'nl',
+
   /** Flemish */
   NlBe = 'nl_BE',
+
   /** Dzongkha */
   Dz = 'dz',
+
   /** English */
   En = 'en',
+
   /** Australian English */
   EnAu = 'en_AU',
+
   /** Canadian English */
   EnCa = 'en_CA',
+
   /** British English */
   EnGb = 'en_GB',
+
   /** American English */
   EnUs = 'en_US',
+
   /** Esperanto */
   Eo = 'eo',
+
   /** Estonian */
   Et = 'et',
+
   /** Ewe */
   Ee = 'ee',
+
   /** Faroese */
   Fo = 'fo',
+
   /** Finnish */
   Fi = 'fi',
+
   /** French */
   Fr = 'fr',
+
   /** Canadian French */
   FrCa = 'fr_CA',
+
   /** Swiss French */
   FrCh = 'fr_CH',
+
   /** Fulah */
   Ff = 'ff',
+
   /** Galician */
   Gl = 'gl',
+
   /** Ganda */
   Lg = 'lg',
+
   /** Georgian */
   Ka = 'ka',
+
   /** German */
   De = 'de',
+
   /** Austrian German */
   DeAt = 'de_AT',
+
   /** Swiss High German */
   DeCh = 'de_CH',
+
   /** Greek */
   El = 'el',
+
   /** Gujarati */
   Gu = 'gu',
+
   /** Haitian Creole */
   Ht = 'ht',
+
   /** Hausa */
   Ha = 'ha',
+
   /** Hebrew */
   He = 'he',
+
   /** Hindi */
   Hi = 'hi',
+
   /** Hungarian */
   Hu = 'hu',
+
   /** Icelandic */
   Is = 'is',
+
   /** Igbo */
   Ig = 'ig',
+
   /** Indonesian */
   Id = 'id',
+
   /** Interlingua */
   Ia = 'ia',
+
   /** Irish */
   Ga = 'ga',
+
   /** Italian */
   It = 'it',
+
   /** Japanese */
   Ja = 'ja',
+
   /** Javanese */
   Jv = 'jv',
+
   /** Kalaallisut */
   Kl = 'kl',
+
   /** Kannada */
   Kn = 'kn',
+
   /** Kashmiri */
   Ks = 'ks',
+
   /** Kazakh */
   Kk = 'kk',
+
   /** Khmer */
   Km = 'km',
+
   /** Kikuyu */
   Ki = 'ki',
+
   /** Kinyarwanda */
   Rw = 'rw',
+
   /** Korean */
   Ko = 'ko',
+
   /** Kurdish */
   Ku = 'ku',
+
   /** Kyrgyz */
   Ky = 'ky',
+
   /** Lao */
   Lo = 'lo',
+
   /** Latin */
   La = 'la',
+
   /** Latvian */
   Lv = 'lv',
+
   /** Lingala */
   Ln = 'ln',
+
   /** Lithuanian */
   Lt = 'lt',
+
   /** Luba-Katanga */
   Lu = 'lu',
+
   /** Luxembourgish */
   Lb = 'lb',
+
   /** Macedonian */
   Mk = 'mk',
+
   /** Malagasy */
   Mg = 'mg',
+
   /** Malay */
   Ms = 'ms',
+
   /** Malayalam */
   Ml = 'ml',
+
   /** Maltese */
   Mt = 'mt',
+
   /** Manx */
   Gv = 'gv',
+
   /** Maori */
   Mi = 'mi',
+
   /** Marathi */
   Mr = 'mr',
+
   /** Mongolian */
   Mn = 'mn',
+
   /** Nepali */
   Ne = 'ne',
+
   /** North Ndebele */
   Nd = 'nd',
+
   /** Northern Sami */
   Se = 'se',
+
   /** Norwegian Bokmål */
   Nb = 'nb',
+
   /** Norwegian Nynorsk */
   Nn = 'nn',
+
   /** Nyanja */
   Ny = 'ny',
+
   /** Odia */
   Or = 'or',
+
   /** Oromo */
   Om = 'om',
+
   /** Ossetic */
   Os = 'os',
+
   /** Pashto */
   Ps = 'ps',
+
   /** Persian */
   Fa = 'fa',
+
   /** Dari */
   FaAf = 'fa_AF',
+
   /** Polish */
   Pl = 'pl',
+
   /** Portuguese */
   Pt = 'pt',
+
   /** Brazilian Portuguese */
   PtBr = 'pt_BR',
+
   /** European Portuguese */
   PtPt = 'pt_PT',
+
   /** Punjabi */
   Pa = 'pa',
+
   /** Quechua */
   Qu = 'qu',
+
   /** Romanian */
   Ro = 'ro',
+
   /** Moldavian */
   RoMd = 'ro_MD',
+
   /** Romansh */
   Rm = 'rm',
+
   /** Rundi */
   Rn = 'rn',
+
   /** Russian */
   Ru = 'ru',
+
   /** Samoan */
   Sm = 'sm',
+
   /** Sango */
   Sg = 'sg',
+
   /** Sanskrit */
   Sa = 'sa',
+
   /** Scottish Gaelic */
   Gd = 'gd',
+
   /** Serbian */
   Sr = 'sr',
+
   /** Shona */
   Sn = 'sn',
+
   /** Sichuan Yi */
   Ii = 'ii',
+
   /** Sindhi */
   Sd = 'sd',
+
   /** Sinhala */
   Si = 'si',
+
   /** Slovak */
   Sk = 'sk',
+
   /** Slovenian */
   Sl = 'sl',
+
   /** Somali */
   So = 'so',
+
   /** Southern Sotho */
   St = 'st',
+
   /** Spanish */
   Es = 'es',
+
   /** European Spanish */
   EsEs = 'es_ES',
+
   /** Mexican Spanish */
   EsMx = 'es_MX',
+
   /** Sundanese */
   Su = 'su',
+
   /** Swahili */
   Sw = 'sw',
+
   /** Congo Swahili */
   SwCd = 'sw_CD',
+
   /** Swedish */
   Sv = 'sv',
+
   /** Tajik */
   Tg = 'tg',
+
   /** Tamil */
   Ta = 'ta',
+
   /** Tatar */
   Tt = 'tt',
+
   /** Telugu */
   Te = 'te',
+
   /** Thai */
   Th = 'th',
+
   /** Tibetan */
   Bo = 'bo',
+
   /** Tigrinya */
   Ti = 'ti',
+
   /** Tongan */
   To = 'to',
+
   /** Turkish */
   Tr = 'tr',
+
   /** Turkmen */
   Tk = 'tk',
+
   /** Ukrainian */
   Uk = 'uk',
+
   /** Urdu */
   Ur = 'ur',
+
   /** Uyghur */
   Ug = 'ug',
+
   /** Uzbek */
   Uz = 'uz',
+
   /** Vietnamese */
   Vi = 'vi',
+
   /** Volapük */
   Vo = 'vo',
+
   /** Welsh */
   Cy = 'cy',
+
   /** Western Frisian */
   Fy = 'fy',
+
   /** Wolof */
   Wo = 'wo',
+
   /** Xhosa */
   Xh = 'xh',
+
   /** Yiddish */
   Yi = 'yi',
+
   /** Yoruba */
   Yo = 'yo',
+
   /** Zulu */
   Zu = 'zu'
 }
@@ -1453,38 +1769,55 @@ export type MissingPasswordError = ErrorResult & {
 
 export type Mutation = {
   __typename?: 'Mutation';
+
   /** Adds an item to the order. If custom fields are defined on the OrderLine entity, a third argument 'customFields' will be available. */
   addItemToOrder: UpdateOrderItemsResult;
+
   /** Remove an OrderLine from the Order */
   removeOrderLine: RemoveOrderItemsResult;
+
   /** Remove all OrderLine from the Order */
   removeAllOrderLines: RemoveOrderItemsResult;
+
   /** Adjusts an OrderLine. If custom fields are defined on the OrderLine entity, a third argument 'customFields' of type `OrderLineCustomFieldsInput` will be available. */
   adjustOrderLine: UpdateOrderItemsResult;
+
   /** Applies the given coupon code to the active Order */
   applyCouponCode: ApplyCouponCodeResult;
+
   /** Removes the given coupon code from the active Order */
   removeCouponCode?: Maybe<Order>;
+
   /** Transitions an Order to a new state. Valid next states can be found by querying `nextOrderStates` */
   transitionOrderToState?: Maybe<TransitionOrderToStateResult>;
+
   /** Sets the shipping address for this order */
   setOrderShippingAddress: ActiveOrderResult;
+
   /** Sets the billing address for this order */
   setOrderBillingAddress: ActiveOrderResult;
+
   /** Allows any custom fields to be set for the active order */
   setOrderCustomFields: ActiveOrderResult;
+
   /** Sets the shipping method by id, which can be obtained with the `eligibleShippingMethods` query */
   setOrderShippingMethod: SetOrderShippingMethodResult;
+
   /** Add a Payment to the Order */
   addPaymentToOrder: AddPaymentToOrderResult;
+
   /** Set the Customer for the Order. Required only if the Customer is not currently logged in */
   setCustomerForOrder: SetCustomerForOrderResult;
+
   /** Authenticates the user using the native authentication strategy. This mutation is an alias for `authenticate({ native: { ... }})` */
   login: NativeAuthenticationResult;
+
   /** Authenticates the user using a named authentication strategy */
   authenticate: AuthenticationResult;
+
   /** End the current authenticated session */
   logout: Success;
+
   /**
    * Register a Customer account with the given credentials. There are three possible registration flows:
    *
@@ -1502,16 +1835,22 @@ export type Mutation = {
    * 3. The Customer _must_ be registered _with_ a password. No further action is needed - the Customer is able to authenticate immediately.
    */
   registerCustomerAccount: RegisterCustomerAccountResult;
+
   /** Regenerate and send a verification token for a new Customer registration. Only applicable if `authOptions.requireVerification` is set to true. */
   refreshCustomerVerification: RefreshCustomerVerificationResult;
+
   /** Update an existing Customer */
   updateCustomer: Customer;
+
   /** Create a new Customer Address */
   createCustomerAddress: Address;
+
   /** Update an existing Address */
   updateCustomerAddress: Address;
+
   /** Delete an existing Address */
   deleteCustomerAddress: Success;
+
   /**
    * Verify a Customer email address with the token sent to that address. Only applicable if `authOptions.requireVerification` is set to true.
    *
@@ -1519,8 +1858,10 @@ export type Mutation = {
    * provided here.
    */
   verifyCustomerAccount: VerifyCustomerAccountResult;
+
   /** Update the password of the active Customer */
   updateCustomerPassword: UpdateCustomerPasswordResult;
+
   /**
    * Request to update the emailAddress of the active Customer. If `authOptions.requireVerification` is enabled
    * (as is the default), then the `identifierChangeToken` will be assigned to the current User and
@@ -1528,79 +1869,69 @@ export type Mutation = {
    * that verification token to the Customer, which is then used to verify the change of email address.
    */
   requestUpdateCustomerEmailAddress: RequestUpdateCustomerEmailAddressResult;
+
   /**
    * Confirm the update of the emailAddress with the provided token, which has been generated by the
    * `requestUpdateCustomerEmailAddress` mutation.
    */
   updateCustomerEmailAddress: UpdateCustomerEmailAddressResult;
+
   /** Requests a password reset email to be sent */
   requestPasswordReset?: Maybe<RequestPasswordResetResult>;
+
   /** Resets a Customer's password based on the provided token */
   resetPassword: ResetPasswordResult;
 };
-
 
 export type MutationAddItemToOrderArgs = {
   productVariantId: Scalars['ID'];
   quantity: Scalars['Int'];
 };
 
-
 export type MutationRemoveOrderLineArgs = {
   orderLineId: Scalars['ID'];
 };
-
 
 export type MutationAdjustOrderLineArgs = {
   orderLineId: Scalars['ID'];
   quantity: Scalars['Int'];
 };
 
-
 export type MutationApplyCouponCodeArgs = {
   couponCode: Scalars['String'];
 };
-
 
 export type MutationRemoveCouponCodeArgs = {
   couponCode: Scalars['String'];
 };
 
-
 export type MutationTransitionOrderToStateArgs = {
   state: Scalars['String'];
 };
-
 
 export type MutationSetOrderShippingAddressArgs = {
   input: CreateAddressInput;
 };
 
-
 export type MutationSetOrderBillingAddressArgs = {
   input: CreateAddressInput;
 };
-
 
 export type MutationSetOrderCustomFieldsArgs = {
   input: UpdateOrderInput;
 };
 
-
 export type MutationSetOrderShippingMethodArgs = {
   shippingMethodId: Scalars['ID'];
 };
-
 
 export type MutationAddPaymentToOrderArgs = {
   input: PaymentInput;
 };
 
-
 export type MutationSetCustomerForOrderArgs = {
   input: CreateCustomerInput;
 };
-
 
 export type MutationLoginArgs = {
   username: Scalars['String'];
@@ -1608,70 +1939,57 @@ export type MutationLoginArgs = {
   rememberMe?: Maybe<Scalars['Boolean']>;
 };
 
-
 export type MutationAuthenticateArgs = {
   input: AuthenticationInput;
   rememberMe?: Maybe<Scalars['Boolean']>;
 };
 
-
 export type MutationRegisterCustomerAccountArgs = {
   input: RegisterCustomerInput;
 };
-
 
 export type MutationRefreshCustomerVerificationArgs = {
   emailAddress: Scalars['String'];
 };
 
-
 export type MutationUpdateCustomerArgs = {
   input: UpdateCustomerInput;
 };
-
 
 export type MutationCreateCustomerAddressArgs = {
   input: CreateAddressInput;
 };
 
-
 export type MutationUpdateCustomerAddressArgs = {
   input: UpdateAddressInput;
 };
 
-
 export type MutationDeleteCustomerAddressArgs = {
   id: Scalars['ID'];
 };
-
 
 export type MutationVerifyCustomerAccountArgs = {
   token: Scalars['String'];
   password?: Maybe<Scalars['String']>;
 };
 
-
 export type MutationUpdateCustomerPasswordArgs = {
   currentPassword: Scalars['String'];
   newPassword: Scalars['String'];
 };
-
 
 export type MutationRequestUpdateCustomerEmailAddressArgs = {
   password: Scalars['String'];
   newEmailAddress: Scalars['String'];
 };
 
-
 export type MutationUpdateCustomerEmailAddressArgs = {
   token: Scalars['String'];
 };
 
-
 export type MutationRequestPasswordResetArgs = {
   emailAddress: Scalars['String'];
 };
-
 
 export type MutationResetPasswordArgs = {
   token: Scalars['String'];
@@ -1742,20 +2060,24 @@ export type Order = Node & {
   id: Scalars['ID'];
   createdAt: Scalars['DateTime'];
   updatedAt: Scalars['DateTime'];
+
   /**
    * The date & time that the Order was placed, i.e. the Customer
    * completed the checkout and the Order is no longer "active"
    */
   orderPlacedAt?: Maybe<Scalars['DateTime']>;
+
   /** A unique code for the Order */
   code: Scalars['String'];
   state: Scalars['String'];
+
   /** An order is active as long as the payment process has not been completed */
   active: Scalars['Boolean'];
   customer?: Maybe<Customer>;
   shippingAddress?: Maybe<OrderAddress>;
   billingAddress?: Maybe<OrderAddress>;
   lines: Array<OrderLine>;
+
   /**
    * Surcharges are arbitrary modifications to the Order total which are neither
    * ProductVariants nor discounts resulting from applied Promotions. For example,
@@ -1764,13 +2086,16 @@ export type Order = Node & {
    */
   surcharges: Array<Surcharge>;
   discounts: Array<Discount>;
+
   /** An array of all coupon codes applied to the Order */
   couponCodes: Array<Scalars['String']>;
+
   /** Promotions applied to the order. Only gets populated after the payment process has completed. */
   promotions: Array<Promotion>;
   payments?: Maybe<Array<Payment>>;
   fulfillments?: Maybe<Array<Fulfillment>>;
   totalQuantity: Scalars['Int'];
+
   /**
    * The subTotal is the total of all OrderLines in the Order. This figure also includes any Order-level
    * discounts which have been prorated (proportionally distributed) amongst the OrderItems.
@@ -1778,22 +2103,25 @@ export type Order = Node & {
    * sum of `OrderLine.discountedLinePrice` values.
    */
   subTotal: Scalars['Int'];
+
   /** Same as subTotal, but inclusive of tax */
   subTotalWithTax: Scalars['Int'];
   currencyCode: CurrencyCode;
   shippingLines: Array<ShippingLine>;
   shipping: Scalars['Int'];
   shippingWithTax: Scalars['Int'];
+
   /** Equal to subTotal plus shipping */
   total: Scalars['Int'];
+
   /** The final payable amount. Equal to subTotalWithTax plus shippingWithTax */
   totalWithTax: Scalars['Int'];
+
   /** A summary of the taxes being applied to this Order */
   taxSummary: Array<OrderTaxSummary>;
   history: HistoryEntryList;
   customFields?: Maybe<Scalars['JSON']>;
 };
-
 
 export type OrderHistoryArgs = {
   options?: Maybe<HistoryEntryListOptions>;
@@ -1837,10 +2165,13 @@ export type OrderItem = Node & {
   createdAt: Scalars['DateTime'];
   updatedAt: Scalars['DateTime'];
   cancelled: Scalars['Boolean'];
+
   /** The price of a single unit, excluding tax and discounts */
   unitPrice: Scalars['Int'];
+
   /** The price of a single unit, including tax but excluding discounts */
   unitPriceWithTax: Scalars['Int'];
+
   /**
    * The price of a single unit including discounts, excluding tax.
    *
@@ -1850,14 +2181,17 @@ export type OrderItem = Node & {
    * about the internal handling of distributed Order-level discounts.
    */
   discountedUnitPrice: Scalars['Int'];
+
   /** The price of a single unit including discounts and tax */
   discountedUnitPriceWithTax: Scalars['Int'];
+
   /**
    * The actual unit price, taking into account both item discounts _and_ prorated (proportially-distributed)
    * Order-level discounts. This value is the true economic value of the OrderItem, and is used in tax
    * and refund calculations.
    */
   proratedUnitPrice: Scalars['Int'];
+
   /** The proratedUnitPrice including tax */
   proratedUnitPriceWithTax: Scalars['Int'];
   unitTax: Scalars['Int'];
@@ -1883,14 +2217,19 @@ export type OrderLine = Node & {
   updatedAt: Scalars['DateTime'];
   productVariant: ProductVariant;
   featuredAsset?: Maybe<Asset>;
+
   /** The price of a single unit, excluding tax and discounts */
   unitPrice: Scalars['Int'];
+
   /** The price of a single unit, including tax but excluding discounts */
   unitPriceWithTax: Scalars['Int'];
+
   /** Non-zero if the unitPrice has changed since it was initially added to Order */
   unitPriceChangeSinceAdded: Scalars['Int'];
+
   /** Non-zero if the unitPriceWithTax has changed since it was initially added to Order */
   unitPriceWithTaxChangeSinceAdded: Scalars['Int'];
+
   /**
    * The price of a single unit including discounts, excluding tax.
    *
@@ -1900,35 +2239,45 @@ export type OrderLine = Node & {
    * about the internal handling of distributed Order-level discounts.
    */
   discountedUnitPrice: Scalars['Int'];
+
   /** The price of a single unit including discounts and tax */
   discountedUnitPriceWithTax: Scalars['Int'];
+
   /**
    * The actual unit price, taking into account both item discounts _and_ prorated (proportially-distributed)
    * Order-level discounts. This value is the true economic value of the OrderItem, and is used in tax
    * and refund calculations.
    */
   proratedUnitPrice: Scalars['Int'];
+
   /** The proratedUnitPrice including tax */
   proratedUnitPriceWithTax: Scalars['Int'];
   quantity: Scalars['Int'];
   items: Array<OrderItem>;
   taxRate: Scalars['Float'];
+
   /** The total price of the line excluding tax and discounts. */
   linePrice: Scalars['Int'];
+
   /** The total price of the line including tax bit excluding discounts. */
   linePriceWithTax: Scalars['Int'];
+
   /** The price of the line including discounts, excluding tax */
   discountedLinePrice: Scalars['Int'];
+
   /** The price of the line including discounts and tax */
   discountedLinePriceWithTax: Scalars['Int'];
+
   /**
    * The actual line price, taking into account both item discounts _and_ prorated (proportially-distributed)
    * Order-level discounts. This value is the true economic value of the OrderLine, and is used in tax
    * and refund calculations.
    */
   proratedLinePrice: Scalars['Int'];
+
   /** The proratedLinePrice including tax */
   proratedLinePriceWithTax: Scalars['Int'];
+
   /** The total tax on this line */
   lineTax: Scalars['Int'];
   discounts: Array<Discount>;
@@ -1996,12 +2345,16 @@ export type OrderStateTransitionError = ErrorResult & {
  */
 export type OrderTaxSummary = {
   __typename?: 'OrderTaxSummary';
+
   /** A description of this tax */
   description: Scalars['String'];
+
   /** The taxRate as a percentage */
   taxRate: Scalars['Float'];
+
   /** The total net price or OrderItems to which this taxRate applies */
   taxBase: Scalars['Int'];
+
   /** The total tax being applied to the Order at this taxRate */
   taxTotal: Scalars['Int'];
 };
@@ -2070,8 +2423,10 @@ export type PaymentFailedError = ErrorResult & {
 
 /** Passed as input to the `addPaymentToOrder` mutation. */
 export type PaymentInput = {
+
   /** This field should correspond to the `code` property of a PaymentMethodHandler. */
   method: Scalars['String'];
+
   /**
    * This field should contain arbitrary data passed to the specified PaymentMethodHandler's `createPayment()` method
    * as the "metadata" argument. For example, it could contain an ID for the payment and other
@@ -2098,174 +2453,259 @@ export type PaymentMethodQuote = {
  * @docsCategory common
  */
 export enum Permission {
+
   /** Authenticated means simply that the user is logged in */
   Authenticated = 'Authenticated',
+
   /** SuperAdmin has unrestricted access to all operations */
   SuperAdmin = 'SuperAdmin',
+
   /** Owner means the user owns this entity, e.g. a Customer's own Order */
   Owner = 'Owner',
+
   /** Public means any unauthenticated user may perform the operation */
   Public = 'Public',
+
   /** Grants permission to update GlobalSettings */
   UpdateGlobalSettings = 'UpdateGlobalSettings',
+
   /** Grants permission to create Products, Facets, Assets, Collections */
   CreateCatalog = 'CreateCatalog',
+
   /** Grants permission to read Products, Facets, Assets, Collections */
   ReadCatalog = 'ReadCatalog',
+
   /** Grants permission to update Products, Facets, Assets, Collections */
   UpdateCatalog = 'UpdateCatalog',
+
   /** Grants permission to delete Products, Facets, Assets, Collections */
   DeleteCatalog = 'DeleteCatalog',
+
   /** Grants permission to create PaymentMethods, ShippingMethods, TaxCategories, TaxRates, Zones, Countries, System & GlobalSettings */
   CreateSettings = 'CreateSettings',
+
   /** Grants permission to read PaymentMethods, ShippingMethods, TaxCategories, TaxRates, Zones, Countries, System & GlobalSettings */
   ReadSettings = 'ReadSettings',
+
   /** Grants permission to update PaymentMethods, ShippingMethods, TaxCategories, TaxRates, Zones, Countries, System & GlobalSettings */
   UpdateSettings = 'UpdateSettings',
+
   /** Grants permission to delete PaymentMethods, ShippingMethods, TaxCategories, TaxRates, Zones, Countries, System & GlobalSettings */
   DeleteSettings = 'DeleteSettings',
+
   /** Grants permission to create Administrator */
   CreateAdministrator = 'CreateAdministrator',
+
   /** Grants permission to read Administrator */
   ReadAdministrator = 'ReadAdministrator',
+
   /** Grants permission to update Administrator */
   UpdateAdministrator = 'UpdateAdministrator',
+
   /** Grants permission to delete Administrator */
   DeleteAdministrator = 'DeleteAdministrator',
+
   /** Grants permission to create Asset */
   CreateAsset = 'CreateAsset',
+
   /** Grants permission to read Asset */
   ReadAsset = 'ReadAsset',
+
   /** Grants permission to update Asset */
   UpdateAsset = 'UpdateAsset',
+
   /** Grants permission to delete Asset */
   DeleteAsset = 'DeleteAsset',
+
   /** Grants permission to create Channel */
   CreateChannel = 'CreateChannel',
+
   /** Grants permission to read Channel */
   ReadChannel = 'ReadChannel',
+
   /** Grants permission to update Channel */
   UpdateChannel = 'UpdateChannel',
+
   /** Grants permission to delete Channel */
   DeleteChannel = 'DeleteChannel',
+
   /** Grants permission to create Collection */
   CreateCollection = 'CreateCollection',
+
   /** Grants permission to read Collection */
   ReadCollection = 'ReadCollection',
+
   /** Grants permission to update Collection */
   UpdateCollection = 'UpdateCollection',
+
   /** Grants permission to delete Collection */
   DeleteCollection = 'DeleteCollection',
+
   /** Grants permission to create Country */
   CreateCountry = 'CreateCountry',
+
   /** Grants permission to read Country */
   ReadCountry = 'ReadCountry',
+
   /** Grants permission to update Country */
   UpdateCountry = 'UpdateCountry',
+
   /** Grants permission to delete Country */
   DeleteCountry = 'DeleteCountry',
+
   /** Grants permission to create Customer */
   CreateCustomer = 'CreateCustomer',
+
   /** Grants permission to read Customer */
   ReadCustomer = 'ReadCustomer',
+
   /** Grants permission to update Customer */
   UpdateCustomer = 'UpdateCustomer',
+
   /** Grants permission to delete Customer */
   DeleteCustomer = 'DeleteCustomer',
+
   /** Grants permission to create CustomerGroup */
   CreateCustomerGroup = 'CreateCustomerGroup',
+
   /** Grants permission to read CustomerGroup */
   ReadCustomerGroup = 'ReadCustomerGroup',
+
   /** Grants permission to update CustomerGroup */
   UpdateCustomerGroup = 'UpdateCustomerGroup',
+
   /** Grants permission to delete CustomerGroup */
   DeleteCustomerGroup = 'DeleteCustomerGroup',
+
   /** Grants permission to create Facet */
   CreateFacet = 'CreateFacet',
+
   /** Grants permission to read Facet */
   ReadFacet = 'ReadFacet',
+
   /** Grants permission to update Facet */
   UpdateFacet = 'UpdateFacet',
+
   /** Grants permission to delete Facet */
   DeleteFacet = 'DeleteFacet',
+
   /** Grants permission to create Order */
   CreateOrder = 'CreateOrder',
+
   /** Grants permission to read Order */
   ReadOrder = 'ReadOrder',
+
   /** Grants permission to update Order */
   UpdateOrder = 'UpdateOrder',
+
   /** Grants permission to delete Order */
   DeleteOrder = 'DeleteOrder',
+
   /** Grants permission to create PaymentMethod */
   CreatePaymentMethod = 'CreatePaymentMethod',
+
   /** Grants permission to read PaymentMethod */
   ReadPaymentMethod = 'ReadPaymentMethod',
+
   /** Grants permission to update PaymentMethod */
   UpdatePaymentMethod = 'UpdatePaymentMethod',
+
   /** Grants permission to delete PaymentMethod */
   DeletePaymentMethod = 'DeletePaymentMethod',
+
   /** Grants permission to create Product */
   CreateProduct = 'CreateProduct',
+
   /** Grants permission to read Product */
   ReadProduct = 'ReadProduct',
+
   /** Grants permission to update Product */
   UpdateProduct = 'UpdateProduct',
+
   /** Grants permission to delete Product */
   DeleteProduct = 'DeleteProduct',
+
   /** Grants permission to create Promotion */
   CreatePromotion = 'CreatePromotion',
+
   /** Grants permission to read Promotion */
   ReadPromotion = 'ReadPromotion',
+
   /** Grants permission to update Promotion */
   UpdatePromotion = 'UpdatePromotion',
+
   /** Grants permission to delete Promotion */
   DeletePromotion = 'DeletePromotion',
+
   /** Grants permission to create ShippingMethod */
   CreateShippingMethod = 'CreateShippingMethod',
+
   /** Grants permission to read ShippingMethod */
   ReadShippingMethod = 'ReadShippingMethod',
+
   /** Grants permission to update ShippingMethod */
   UpdateShippingMethod = 'UpdateShippingMethod',
+
   /** Grants permission to delete ShippingMethod */
   DeleteShippingMethod = 'DeleteShippingMethod',
+
   /** Grants permission to create Tag */
   CreateTag = 'CreateTag',
+
   /** Grants permission to read Tag */
   ReadTag = 'ReadTag',
+
   /** Grants permission to update Tag */
   UpdateTag = 'UpdateTag',
+
   /** Grants permission to delete Tag */
   DeleteTag = 'DeleteTag',
+
   /** Grants permission to create TaxCategory */
   CreateTaxCategory = 'CreateTaxCategory',
+
   /** Grants permission to read TaxCategory */
   ReadTaxCategory = 'ReadTaxCategory',
+
   /** Grants permission to update TaxCategory */
   UpdateTaxCategory = 'UpdateTaxCategory',
+
   /** Grants permission to delete TaxCategory */
   DeleteTaxCategory = 'DeleteTaxCategory',
+
   /** Grants permission to create TaxRate */
   CreateTaxRate = 'CreateTaxRate',
+
   /** Grants permission to read TaxRate */
   ReadTaxRate = 'ReadTaxRate',
+
   /** Grants permission to update TaxRate */
   UpdateTaxRate = 'UpdateTaxRate',
+
   /** Grants permission to delete TaxRate */
   DeleteTaxRate = 'DeleteTaxRate',
+
   /** Grants permission to create System */
   CreateSystem = 'CreateSystem',
+
   /** Grants permission to read System */
   ReadSystem = 'ReadSystem',
+
   /** Grants permission to update System */
   UpdateSystem = 'UpdateSystem',
+
   /** Grants permission to delete System */
   DeleteSystem = 'DeleteSystem',
+
   /** Grants permission to create Zone */
   CreateZone = 'CreateZone',
+
   /** Grants permission to read Zone */
   ReadZone = 'ReadZone',
+
   /** Grants permission to update Zone */
   UpdateZone = 'UpdateZone',
+
   /** Grants permission to delete Zone */
   DeleteZone = 'DeleteZone'
 }
@@ -2476,35 +2916,47 @@ export type PromotionList = PaginatedList & {
 
 export type Query = {
   __typename?: 'Query';
+
   /** The active Channel */
   activeChannel: Channel;
+
   /** The active Customer */
   activeCustomer?: Maybe<Customer>;
+
   /**
    * The active Order. Will be `null` until an Order is created via `addItemToOrder`. Once an Order reaches the
    * state of `PaymentApproved` or `PaymentSettled`, then that Order is no longer considered "active" and this
    * query will once again return `null`.
    */
   activeOrder?: Maybe<Order>;
+
   /** An array of supported Countries */
   availableCountries: Array<Country>;
+
   /** A list of Collections available to the shop */
   collections: CollectionList;
+
   /** Returns a Collection either by its id or slug. If neither 'id' nor 'slug' is speicified, an error will result. */
   collection?: Maybe<Collection>;
+
   /** Returns a list of eligible shipping methods based on the current active Order */
   eligibleShippingMethods: Array<ShippingMethodQuote>;
+
   /** Returns a list of payment methods and their eligibility based on the current active Order */
   eligiblePaymentMethods: Array<PaymentMethodQuote>;
+
   /** Returns information about the current authenticated User */
   me?: Maybe<CurrentUser>;
+
   /** Returns the possible next states that the activeOrder can transition to */
   nextOrderStates: Array<Scalars['String']>;
+
   /**
    * Returns an Order based on the id. Note that in the Shop API, only orders belonging to the
    * currently-authenticated User may be queried.
    */
   order?: Maybe<Order>;
+
   /**
    * Returns an Order based on the order `code`. For guest Orders (i.e. Orders placed by non-authenticated Customers)
    * this query will only return the Order within 2 hours of the Order being placed. This allows an Order confirmation
@@ -2512,46 +2964,42 @@ export type Query = {
    * general anonymous access to Order data.
    */
   orderByCode?: Maybe<Order>;
+
   /** Get a Product either by id or slug. If neither 'id' nor 'slug' is speicified, an error will result. */
   product?: Maybe<Product>;
+
   /** Get a list of Products */
   products: ProductList;
+
   /** Search Products based on the criteria set by the `SearchInput` */
   search: SearchResponse;
 };
 
-
 export type QueryCollectionsArgs = {
   options?: Maybe<CollectionListOptions>;
 };
-
 
 export type QueryCollectionArgs = {
   id?: Maybe<Scalars['ID']>;
   slug?: Maybe<Scalars['String']>;
 };
 
-
 export type QueryOrderArgs = {
   id: Scalars['ID'];
 };
 
-
 export type QueryOrderByCodeArgs = {
   code: Scalars['String'];
 };
-
 
 export type QueryProductArgs = {
   id?: Maybe<Scalars['ID']>;
   slug?: Maybe<Scalars['String']>;
 };
 
-
 export type QueryProductsArgs = {
   options?: Maybe<ProductListOptions>;
 };
-
 
 export type QuerySearchArgs = {
   input: SearchInput;
@@ -2668,8 +3116,10 @@ export type SearchResult = {
   description: Scalars['String'];
   facetIds: Array<Scalars['ID']>;
   facetValueIds: Array<Scalars['ID']>;
+
   /** An array of ids of the Collections in which this result appears */
   collectionIds: Array<Scalars['ID']>;
+
   /** A relevence score for the result. Differs between database implementations */
   score: Scalars['Float'];
 };
@@ -2732,6 +3182,7 @@ export type ShippingMethodQuote = {
   code: Scalars['String'];
   name: Scalars['String'];
   description: Scalars['String'];
+
   /** Any optional metadata returned by the ShippingCalculator in the ShippingCalculationResult */
   metadata?: Maybe<Scalars['JSON']>;
 };
@@ -2900,7 +3351,6 @@ export type UpdateOrderInput = {
 };
 
 export type UpdateOrderItemsResult = Order | OrderModificationError | OrderLimitError | NegativeQuantityError | InsufficientStockError;
-
 
 export type User = Node & {
   __typename?: 'User';
