+++ packages/api-client/src/types/context.ts
@@ -1,5 +1,5 @@
-import { ApiClientMethods, IntegrationContext } from "@vue-storefront/core";
-import { VendureApiMethods } from "./API";
-import { ClientInstance, Config } from "./setup";
+import { ApiClientMethods, IntegrationContext } from '@vue-storefront/core';
+import { VendureApiMethods } from './API';
+import { ClientInstance, Config } from './setup';
 
 export type Context = IntegrationContext<ClientInstance, Config, ApiClientMethods<VendureApiMethods>>;
